+++ test/plugins/Collateral.test.ts
@@ -48,7 +48,7 @@ import {
 
 const createFixtureLoader = waffle.createFixtureLoader
 
-const DEFAULT_THRESHOLD = fp('0.05') // 5%
+const DEFAULT_THRESHOLD = fp('0.01') // 1%
 const DELAY_UNTIL_DEFAULT = bn('86400') // 24h
 
 const describeGas = process.env.REPORT_GAS ? describe : describe.skip