+++ test/scenario/cWBTC.test.ts
@@ -31,7 +31,7 @@ import {
 } from '../fixtures'
 import { expectPrice } from '../utils/oracles'
 
-const DEFAULT_THRESHOLD = fp('0.05') // 5%
+const DEFAULT_THRESHOLD = fp('0.01') // 1%
 const DELAY_UNTIL_DEFAULT = bn('86400') // 24h
 
 const createFixtureLoader = waffle.createFixtureLoader